addressing feedback from reviewer

+++ src/processor.cpp
@@ -4,10 +4,10 @@
 // TODO: Return the aggregate CPU utilization
 float Processor::Utilization() 
 {
-    long current_cpu_usage_active = LinuxParser::ActiveJiffies();
-    long current_cpu_usage_idle = LinuxParser::IdleJiffies();
-    long cpu_usage_active_delta = current_cpu_usage_active - previous_cpu_usage_active_; 
-    long cpu_usage_idle_delta = current_cpu_usage_idle - previous_cpu_usage_idle_; 
+    float current_cpu_usage_active = static_cast<float>(LinuxParser::ActiveJiffies());
+    float current_cpu_usage_idle = static_cast<float>(LinuxParser::IdleJiffies());
+    float cpu_usage_active_delta = current_cpu_usage_active - previous_cpu_usage_active_; 
+    float cpu_usage_idle_delta = current_cpu_usage_idle - previous_cpu_usage_idle_; 
     previous_cpu_usage_active_ = current_cpu_usage_active;
     previous_cpu_usage_idle_ = current_cpu_usage_idle;
